+++ todo-app/App.js
@@ -1,90 +1,22 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, TextInput, Button, FlatList, StyleSheet } from 'react-native';
-import axios from 'axios';
-import uuid from 'react-native-uuid'; // ✅ use this instead of 'uuid'
+// App.js
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-const API_URL = 'http://172.24.82.173:3000/todos'; // your local IP
+import SignupScreen from './SignupScreen';
+import LoginScreen from './Loginscreen';
+import TodoScreen from './TodoScreen'; 
+
+const Stack = createNativeStackNavigator();
 
 export default function App() {
-  const [task, setTask] = useState('');
-  const [todos, setTodos] = useState([]);
-
-  const fetchTodos = async () => {
-    try {
-      const res = await axios.get(API_URL);
-      setTodos(res.data);
-    } catch (err) {
-      console.error('Failed to fetch tasks:', err.message);
-    }
-  };
-
-  useEffect(() => {
-    fetchTodos();
-  }, []);
-
-  const addTodo = async () => {
-    if (!task.trim()) return;
-
-    try {
-      const newTodo = { id: uuid.v4(), task }; // ✅ generate UUID
-      await axios.post(API_URL, newTodo);
-      setTask('');
-      fetchTodos();
-    } catch (err) {
-      console.error('Failed to add task:', err.message);
-    }
-  };
-
-  const deleteTodo = async (id) => {
-    try {
-      await axios.delete(`${API_URL}/${id}`);
-      fetchTodos();
-    } catch (err) {
-      console.error('Failed to delete task:', err.message);
-    }
-  };
-
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>To-Do List</Text>
-      <TextInput
-        placeholder="Enter task"
-        value={task}
-        onChangeText={setTask}
-        style={styles.input}
-      />
-      <Button title="Add Task" onPress={addTodo} />
-      <FlatList
-        data={todos}
-        keyExtractor={item => item.id.toString()}
-        renderItem={({ item }) => (
-          <View style={styles.item}>
-            <Text>{item.task}</Text>
-            <Button title="Delete" onPress={() => deleteTodo(item.id)} />
-          </View>
-        )}
-      />
-    </View>
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="Signup">
+        <Stack.Screen name="Signup" component={SignupScreen} />
+        <Stack.Screen name="Login" component={LoginScreen} />
+        <Stack.Screen name="Todo" component={TodoScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: { padding: 30, marginTop: 40 },
-  title: { fontSize: 24, fontWeight: 'bold', marginBottom: 10 },
-  input: {
-    borderColor: '#ccc',
-    borderWidth: 1,
-    marginBottom: 10,
-    padding: 8,
-    borderRadius: 5,
-  },
-  item: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginVertical: 5,
-    paddingVertical: 5,
-    borderBottomColor: '#eee',
-    borderBottomWidth: 1,
-  },
-});
