+++ todo-app/Loginscreen.js
@@ -1,0 +1,39 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, Button, StyleSheet } from 'react-native';
+
+export default function LoginScreen({ navigation }) {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleLogin = () => {
+    // Simulate login
+    if (email && password) {
+      navigation.navigate('Todo');
+    } else {
+      alert('Please enter email and password');
+    }
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Login</Text>
+      <TextInput placeholder="Email" onChangeText={setEmail} style={styles.input} />
+      <TextInput placeholder="Password" secureTextEntry onChangeText={setPassword} style={styles.input} />
+      <Button title="Login" onPress={handleLogin} />
+      <Text onPress={() => navigation.navigate('Signup')} style={styles.link}>Don't have an account? Sign up</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: { padding: 20, marginTop: 50 },
+  title: { fontSize: 24, marginBottom: 20 },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    marginBottom: 15,
+    padding: 10,
+    borderRadius: 5,
+  },
+  link: { marginTop: 10, color: 'blue', textAlign: 'center' },
+});
