+++ todo-app/SignupScreen.js
@@ -1,0 +1,38 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, Button, StyleSheet } from 'react-native';
+
+export default function SignupScreen({ navigation }) {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleSignup = () => {
+    if (email && password) {
+      navigation.navigate('Todo'); // Simulate account creation
+    } else {
+      alert('Please fill all fields');
+    }
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Sign Up</Text>
+      <TextInput placeholder="Email" onChangeText={setEmail} style={styles.input} />
+      <TextInput placeholder="Password" secureTextEntry onChangeText={setPassword} style={styles.input} />
+      <Button title="Sign Up" onPress={handleSignup} />
+      <Text onPress={() => navigation.navigate('Login')} style={styles.link}>Already have an account? Log in</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: { padding: 20, marginTop: 50 },
+  title: { fontSize: 24, marginBottom: 20 },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    marginBottom: 15,
+    padding: 10,
+    borderRadius: 5,
+  },
+  link: { marginTop: 10, color: 'blue', textAlign: 'center' },
+});
